Intro to Jackson ObjectMapper/src/main/java/com/metaxiii/fr/conversion/SerializationDeserialization.java
package com.metaxiii.fr.conversion;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.DeserializationFeature;
import com.fasterxml.jackson.databind.JsonNode;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.metaxiii.fr.model.Car;
import lombok.extern.slf4j.Slf4j;

@Slf4j
public class SerializationDeserialization {

  public Car processWithUnrecognizedPropertyException(String json)
    throws JsonProcessingException {
    ObjectMapper objectMapper = new ObjectMapper();
    return objectMapper.readValue(json, Car.class);
  }

  public Car process(String json) throws JsonProcessingException {
    ObjectMapper objectMapper = new ObjectMapper();
    objectMapper.configure(
      DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES,
      false
    );
    return objectMapper.readValue(json, Car.class);
  }

  public JsonNode processJsonNode(String json) throws JsonProcessingException {
    ObjectMapper objectMapper = new ObjectMapper();
    return objectMapper.readTree(json);
  }
}
